penyempurnaannnn

+++ SewaMotor_Pterpan/FormUtama.vb
@@ -89,11 +89,16 @@
         Me.Close()
     End Sub
     Private Sub btnDelete_Click(sender As Object, e As EventArgs) Handles btnDelete.Click
-        idmotor = dgvMotor.Item(0, dgvMotor.CurrentRow.Index).Value
-        motor.isiDataTable("DELETE FROM Motor WHERE id_motor=" & idmotor, "Berhasil Delete")
-        Dim utama As New FormUtama
-        utama.Show()
-        Me.Close()
+        If dgvMotor.Item(5, dgvMotor.CurrentRow.Index).Value = 0 Then
+            idmotor = dgvMotor.Item(0, dgvMotor.CurrentRow.Index).Value
+            motor.isiDataTable("DELETE FROM Motor WHERE id_motor=" & idmotor, "Berhasil Delete")
+            Dim utama As New FormUtama
+            utama.Show()
+            Me.Close()
+        Else
+            MsgBox("Motor sedang dipinjam")
+        End If
+
     End Sub
     Private Sub btnPengembalian_Click(sender As Object, e As EventArgs) Handles btnPengembalian.Click
         Dim kembali As New FormKembali
